Added VTM16 DAQ Prototype as ran on car Fall Break 2015
This MPLAB X project is the current working state of the VTM Data Aq
system. So far features include: CAN parsing of accelerometer data and
MoTeC outputs on separate CAN busses, Logging to a USB stick as csv
(BTN1 and BTN2 presses required to start and stop logging), 2 analog
inputs are successfully read and one is recorded to the USB csv log.
Others not mentioned